fixed the calculation of locations in templates
So now when there's an error in the compilation of a template, it's
much more likely that the location in the error-message will be
accurate.

Though, actually, since strings/quoted-strings don't have locations
for the contained content, but only for the entire token, the location
will still be off by a little.  Just not completley bonkers wrong.